CAM: Job - Fix Cycle Time
